win screen + added game over
Made it so that when the score reaches a certain point one wins and the win screen appears. Tried to connect the start screen but lost patience. Made it so that the timer stops when you either win or lose cause it was running in the background. Tried adding a second obstacle but found it too hard.